# Conflicts:
#	data/battle_scripts_1.s
#	include/config/pokemon.h
#	src/birch_pc.c

+++ include/birch_pc.h
@@ -1,6 +1,6 @@
 #ifndef GUARD_BIRCH_PC_H
 #define GUARD_BIRCH_PC_H
 
-const u8 *GetPokedexRatingText(u16 count);
+const u8 *GetPokedexRatingText(u32 count);
 
 #endif // GUARD_BIRCH_PC_H
